ConvertTestCases doesn't belong in FilteringTestUtils
